python/pyproject.toml
[tool.poetry]
name = "weather-console"
version = "0.1.0"
description = ""
authors = ["Juan Diego <juandiego500.jdg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
